Revert "refactor: deduplicate max effective square size (#3294)"
This reverts commit e091ac4004645e986810c9c79e5ee93ce52eef86.
